created second fragmnet
