+++ app/src/main/java/com/mikhail/githubapi/MainActivity.java
@@ -16,6 +16,7 @@ import android.widget.Toast;
 
 import com.mikhail.githubapi.adapter.RepositoryAdapter;
 import com.mikhail.githubapi.fragment.ContributorFragment;
+import com.mikhail.githubapi.fragment.RepositoryFragment;
 import com.mikhail.githubapi.model.Items;
 import com.mikhail.githubapi.model.Repo;
 import com.mikhail.githubapi.provider.GitHubAPIService;
@@ -39,15 +40,14 @@ public class MainActivity extends AppCompatActivity {
     protected RecyclerView recyclerView;
     private RepositoryAdapter repositoryAdapter;
     private List<Items> gitHubData;
-    private SwipeRefreshLayout swipeContainer;
     public Context context;
     public static final String REPOS_DATE = "created:>=";
     public static final String REPOS_RATING = "star";
-    public static final String REPOS_SORT= "desc";
+    public static final String REPOS_SORT = "desc";
     public static final String TAG = "MainActivity";
     private FragmentManager fragmentManager;
-
-
+    private ContributorFragment contributorFragment;
+    private RepositoryFragment repositoryFragment;
 
 
     @Override
@@ -55,135 +55,17 @@ public class MainActivity extends AppCompatActivity {
         super.onCreate(savedInstanceState);
         setContentView(R.layout.activity_main);
         context = getApplicationContext();
-        swipeContainer = (SwipeRefreshLayout) findViewById(R.id.swipeContainer);
         fragmentManager = getSupportFragmentManager();
+        contributorFragment = new ContributorFragment();
+        repositoryFragment = new RepositoryFragment();
 
-        checkNetwork();
-        setPullRefresh();
-        reposApiCall();
-        initRecyclerView();
-        reposClickListener();
-    }
-
-    /**
-     * set recycler view adapter
-     */
-    private void initRecyclerView() {
-
-        gitHubData = new ArrayList<>();
-        repositoryAdapter = new RepositoryAdapter(gitHubData);
-        recyclerView = (RecyclerView) findViewById(R.id.my_recycler_view);
-        recyclerView.setLayoutManager(new LinearLayoutManager(this));
-        recyclerView.setAdapter(repositoryAdapter);
-    }
-
-    private void reposApiCall() {
-        GitHubAPIService.GitHubRx gitHub = GitHubAPIService.createRx();
-
-        Observable<Response<Repo>> observable =
-                gitHub.repositories(REPOS_DATE + AppUtils.getLastWeekDate(),
-                REPOS_RATING, REPOS_SORT);
-
-        observable.observeOn(AndroidSchedulers.mainThread()).
-                subscribeOn(Schedulers.io()).
-                observeOn(AndroidSchedulers.mainThread()).
-                subscribe(new Subscriber<Response<Repo>>() {
-                    @Override
-                    public void onCompleted() {
-
-                    }
-
-                    @Override
-                    public void onError(Throwable e) {
-                        Log.d(TAG, "Call failed!");
-
-                    }
-
-                    @Override
-                    public void onNext(Response<Repo> repositories) {
-                        Log.d(TAG, "Call success!");
-
-                        callSuccess(repositories);
-
-                    }
-                });
-    }
-
-    private void callSuccess(Response<Repo> repositories){
-
-        gitHubData = repositories.body().getItems();
-        repositoryAdapter = new RepositoryAdapter(gitHubData);
-        recyclerView.setAdapter(repositoryAdapter);
-        repositoryAdapter.notifyDataSetChanged();
-        swipeContainer.setRefreshing(false);
-    }
-
-    /**
-     * check network connection
-     */
-    private void checkNetwork() {
-
-        if (!isConnected(this.context)) {
-
-            new Timer().schedule(new TimerTask() {
-                @Override
-                public void run() {
-                    Intent intent = new Intent(Settings.ACTION_SETTINGS);
-                    startActivity(intent);
-                }
-            }, 1600);
-
-            Toast.makeText(MainActivity.this, R.string.no_internet, Toast.LENGTH_LONG).show();
-            swipeContainer.setRefreshing(false);
-        } else {
-            setPullRefresh();
-        }
-    }
-
-    private void setPullRefresh() {
-        swipeContainer.setOnRefreshListener(new SwipeRefreshLayout.OnRefreshListener() {
-            @Override
-            public void onRefresh() {
-                if (!isConnected(context)) {
-                    Toast.makeText(MainActivity.this, R.string.no_internet, Toast.LENGTH_LONG).show();
-                    swipeContainer.setRefreshing(false);
-                } else {
-                    reposApiCall();
-                }
-
-            }
-        });
-
-        swipeContainer.setColorSchemeResources(android.R.color.darker_gray,
-                android.R.color.white);
-
-    }
-
-    private void reposClickListener(){
-
-        if (repositoryAdapter != null){
-
-            repositoryAdapter.setOnItemClickListener(new RepositoryAdapter.OnItemClickListener() {
-                @Override
-                public void onItemClick(View itemView, int position) {
-
-                    Items item = gitHubData.get(position);
-
-                    ContributorFragment contributorFragment = new ContributorFragment();
-
-                    Bundle args = new Bundle();
-                    args.putInt("contributorFragment", position);
-                    contributorFragment.setArguments(args);
-                    contributorFragment.setUrl(item.getContributorURL());
-                    FragmentTransaction fragmentTransaction = fragmentManager.beginTransaction();
-                    fragmentTransaction.add(R.id.relative_layout, contributorFragment);
+        FragmentTransaction fragmentTransaction = fragmentManager.beginTransaction();
+                    fragmentTransaction.addToBackStack(null);
+                    fragmentTransaction.replace(R.id.frag_container, repositoryFragment);
                     fragmentTransaction.commit();
-                }
-            });
-        }
-
 
     }
+
 
 }
