+++ app/src/main/java/com/mikhail/githubapi/fragment/ContributorFragment.java
@@ -12,15 +12,11 @@ import android.view.ViewGroup;
 import android.widget.ImageView;
 import android.widget.TextView;
 
-import com.mikhail.githubapi.MainActivity;
 import com.mikhail.githubapi.R;
 import com.mikhail.githubapi.adapter.ContributorAdapter;
-import com.mikhail.githubapi.adapter.RepositoryAdapter;
 import com.mikhail.githubapi.model.Contributor;
 import com.mikhail.githubapi.model.Items;
-import com.mikhail.githubapi.model.Repo;
 import com.mikhail.githubapi.provider.GitHubAPIService;
-import com.mikhail.githubapi.util.AppUtils;
 
 import java.util.ArrayList;
 import java.util.List;
@@ -52,19 +48,20 @@ public class ContributorFragment extends Fragment {
 
         // Hit API to get Contributors for the selected repository if network
 //        getContributors(getArguments().getInt("contributorFragment" ));
+        getContributors();
     }
 
     @Nullable
     @Override
     public View onCreateView(LayoutInflater inflater, ViewGroup container, Bundle savedInstanceState) {
 
-        v = inflater.inflate(R.layout.fragment_layout, container, false);
+        v = inflater.inflate(R.layout.fragment_contributor_layout, container, false);
         contributorName = (TextView) v.findViewById(R.id.contributor_name);
         contribution = (TextView) v.findViewById(R.id.contribution);
         contributorImage = (ImageView) v.findViewById(R.id.img_contributor);
 
 
-        contributorRecyclerView = (RecyclerView) v.findViewById(R.id.contributor_recycler_view);
+        contributorRecyclerView = (RecyclerView) v.findViewById(R.id.recycler_view);
         RecyclerView.LayoutManager contributorLayoutManager = new LinearLayoutManager(getActivity());
         contributorRecyclerView.setLayoutManager(contributorLayoutManager);
 
@@ -112,7 +109,7 @@ public class ContributorFragment extends Fragment {
                     }
                 });
     }
-    
+
 
     public void setUrl(String url) {
         this.url = url;
