+++ app/src/main/res/layout/fragment_recycler_view.xml
@@ -1,3 +1,9 @@
+<android.support.v4.widget.SwipeRefreshLayout
+    android:id="@+id/swipeContainer"
+    xmlns:android="http://schemas.android.com/apk/res/android"
+    android:layout_width="match_parent"
+    android:layout_height="match_parent">
+
 <RelativeLayout xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
@@ -11,9 +17,11 @@
     tools:ignore="Overdraw">
 
     <android.support.v7.widget.RecyclerView
-        android:id="@+id/contributor_recycler_view"
+        android:id="@+id/recycler_view"
         android:layout_width="match_parent"
         android:layout_height="match_parent"
         android:scrollbars="vertical" />
 
 </RelativeLayout>
+
+</android.support.v4.widget.SwipeRefreshLayout>
