+++ app/src/main/res/layout/fragment_repo_layout.xml
@@ -1,3 +1,6 @@
+
+
+
 <android.support.v7.widget.CardView xmlns:android="http://schemas.android.com/apk/res/android"
     android:id="@+id/card_view"
     android:layout_width="match_parent"
